Move tests to BridgeTest and use bridgeBuilder
